Stop leaning on (interactive) for custom args
Since we are using custom logic to prompt anyway, we don't get much by
wiring into (interactive) and we lose the ability to get other args from
interactive, like prefix args.

By filling a var instead of using interactive args, we are free to use
interactive to prompt for more.