Order ido choices instead of providing default
Defaults don't really work so well for ido completions because there's
no way to provide an empty string if you have a list of values, unless
the empty string is one of those values. The way smex handles this is to
just put the default choice at the beginning of the list, giving you a
simple UI and minimal keystrokes. We've followed that here.